Revert file rename, cannot fix git not detecting it properly
