+++ src/types/index.ts
@@ -1,3 +1,3 @@
-export * from "./task-and-batch.js";
+export * from "./task_and_batch.js";
 export * from "./token.js";
 export * from "./types.js";
